databricks_client.py
import logging
from databricks.sdk import WorkspaceClient
from databricks.sdk.service.sql import QueryStatus
import pandas as pd
from typing import Dict, List, Optional
from config import DATABRICKS_HOST, DATABRICKS_TOKEN, DATABRICKS_CATALOG, DATABRICKS_SCHEMA, DATABRICKS_TABLES

logger = logging.getLogger(__name__)

class DatabricksClient:
    def __init__(self):
        """Initialize Databricks client with your credentials"""
        # TODO: Add your Databricks workspace URL and token
        self.client = WorkspaceClient(
            host=DATABRICKS_HOST,
            token=DATABRICKS_TOKEN
        )
        
    def query_vehicle_program_status(self, launch_date: str) -> Dict[str, any]:
        """
        Query all department statuses for a given vehicle program launch date
        
        Args:
            launch_date (str): Date in YYYY-MM-DD format
            
        Returns:
            Dict containing status information from all departments
        """
        try:
            results = {}
            
            # Query Bill of Material status
            bom_status = self._query_bom_status(launch_date)
            results['bill_of_material'] = bom_status
            
            # Query Master Parts List status
            mpl_status = self._query_mpl_status(launch_date)
            results['master_parts_list'] = mpl_status
            
            # Query Material Flow Engineering status
            mfe_status = self._query_mfe_status(launch_date)
            results['material_flow_engineering'] = mfe_status
            
            # Query 4P status
            p4_status = self._query_4p_status(launch_date)
            results['4p'] = p4_status
            
            # Query PPAP status
            ppap_status = self._query_ppap_status(launch_date)
            results['ppap'] = ppap_status
            
            return results
            
        except Exception as e:
            logger.error(f"Error querying vehicle program status: {e}")
            raise
    
    def _query_bom_status(self, launch_date: str) -> Dict:
        """Query Bill of Material status"""
        # TODO: Update this query based on your actual BOM table structure
        query = f"""
        SELECT 
            part_number,
            part_name,
            status,
            completion_percentage,
            last_updated
        FROM {DATABRICKS_CATALOG}.{DATABRICKS_SCHEMA}.{DATABRICKS_TABLES['bill_of_material']}
        WHERE launch_date = '{launch_date}'
        """
        return self._execute_query(query, "BOM")
    
    def _query_mpl_status(self, launch_date: str) -> Dict:
        """Query Master Parts List status"""
        # TODO: Update this query based on your actual MPL table structure
        query = f"""
        SELECT 
            part_id,
            part_description,
            status,
            supplier_info,
            lead_time
        FROM {DATABRICKS_CATALOG}.{DATABRICKS_SCHEMA}.{DATABRICKS_TABLES['master_parts_list']}
        WHERE launch_date = '{launch_date}'
        """
        return self._execute_query(query, "MPL")
    
    def _query_mfe_status(self, launch_date: str) -> Dict:
        """Query Material Flow Engineering status"""
        # TODO: Update this query based on your actual MFE table structure
        query = f"""
        SELECT 
            flow_id,
            process_step,
            status,
            cycle_time,
            efficiency_rating
        FROM {DATABRICKS_CATALOG}.{DATABRICKS_SCHEMA}.{DATABRICKS_TABLES['material_flow_engineering']}
        WHERE launch_date = '{launch_date}'
        """
        return self._execute_query(query, "MFE")
    
    def _query_4p_status(self, launch_date: str) -> Dict:
        """Query 4P status"""
        # TODO: Update this query based on your actual 4P table structure
        query = f"""
        SELECT 
            process_id,
            process_name,
            status,
            people_assigned,
            place_location,
            product_impact
        FROM {DATABRICKS_CATALOG}.{DATABRICKS_SCHEMA}.{DATABRICKS_TABLES['4p']}
        WHERE launch_date = '{launch_date}'
        """
        return self._execute_query(query, "4P")
    
    def _query_ppap_status(self, launch_date: str) -> Dict:
        """Query PPAP status"""
        # TODO: Update this query based on your actual PPAP table structure
        query = f"""
        SELECT 
            ppap_id,
            submission_level,
            status,
            approval_date,
            comments
        FROM {DATABRICKS_CATALOG}.{DATABRICKS_SCHEMA}.{DATABRICKS_TABLES['ppap']}
        WHERE launch_date = '{launch_date}'
        """
        return self._execute_query(query, "PPAP")
    
    def _execute_query(self, query: str, department: str) -> Dict:
        """Execute SQL query and return results"""
        try:
            # TODO: Implement actual query execution based on your Databricks setup
            # This is a placeholder - you'll need to implement the actual query execution
            logger.info(f"Executing query for {department}: {query}")
            
            # Placeholder response - replace with actual query execution
            return {
                'status': 'success',
                'data': [],  # TODO: Replace with actual query results
                'summary': {
                    'total_items': 0,
                    'completed': 0,
                    'pending': 0,
                    'overdue': 0
                }
            }
            
        except Exception as e:
            logger.error(f"Error executing query for {department}: {e}")
            return {
                'status': 'error',
                'error': str(e),
                'data': [],
                'summary': {
                    'total_items': 0,
                    'completed': 0,
                    'pending': 0,
                    'overdue': 0
                }
            }
    
    def create_visualization(self, data: Dict, launch_date: str) -> str:
        """
        Create visualization in Databricks
        
        Args:
            data: Dictionary containing all department data
            launch_date: Launch date for the vehicle program
            
        Returns:
            URL or identifier for the created visualization
        """
        try:
            # TODO: Implement Databricks visualization creation
            # This could involve creating dashboards, charts, or notebooks
            
            logger.info(f"Creating visualization for launch date: {launch_date}")
            
            # Placeholder - implement actual visualization creation
            visualization_url = f"https://your-databricks-workspace.com/visualization/{launch_date}"
            
            return visualization_url
            
        except Exception as e:
            logger.error(f"Error creating visualization: {e}")
            raise 