requirements.txt
slack-bolt==1.23.0
openai==1.3.0
databricks-sdk==0.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pandas==2.1.3
numpy==1.24.3
plotly==5.17.0
python-dotenv==1.0.0
openpyxl==3.1.2
gspread==5.12.0
requests==2.31.0
python-dateutil==2.8.2 